Irisa Morph tech - challenge room
This is just the challenge room to unlock the Irisa Morph tech, and the Outpost quest to get there.

I've been prototyping the Irisa Morph tech using the Crab Morph mod as a starting point to learn from, but I don't want to upload a copied-and-pasted, slightly-modified crab morph to the repo, and will wait until the code for the Irisa Morph tech is complete before pushing those particular assets.